refactor: ARM-242 move useDrag hook to switchInput folder, as this is the only place it's being used.
