+++ module/src/components/switchInput/useDrag.ts
@@ -1,13 +1,15 @@
-import * as React from 'react';
+import * as React from "react";
 
-import { useEventListener } from './useEventListener';
+import { useEventListener } from "../../hooks/useEventListener";
 
 interface IPosition {
   left: number;
   top: number;
 }
 
-function mouseEventToPosition(event: Pick<React.MouseEvent, 'clientX' | 'clientY'>): IPosition {
+function mouseEventToPosition(
+  event: Pick<React.MouseEvent, "clientX" | "clientY">
+): IPosition {
   return { left: event.clientX, top: event.clientY };
 }
 
@@ -19,8 +21,12 @@ export interface IDragReleaseCallbackArgs {
 
 /** Track the position of the cursor after a mouse down event, and fire a callback when the mouse is released */
 export function useDrag(onRelease?: (args: IDragReleaseCallbackArgs) => void) {
-  const [startPosition, setStartPosition] = React.useState<IPosition | undefined>(undefined);
-  const [currentPosition, setCurrentPosition] = React.useState<IPosition | undefined>(undefined);
+  const [startPosition, setStartPosition] = React.useState<
+    IPosition | undefined
+  >(undefined);
+  const [currentPosition, setCurrentPosition] = React.useState<
+    IPosition | undefined
+  >(undefined);
   const [isDragging, setIsDragging] = React.useState(false);
 
   const changePosition = React.useMemo(() => {
@@ -57,15 +63,19 @@ export function useDrag(onRelease?: (args: IDragReleaseCallbackArgs) => void) {
 
   const onMouseUpWindow = React.useCallback(() => {
     if (isDragging) {
-      onRelease?.({ startPosition: startPosition!, currentPosition: currentPosition!, changePosition: changePosition! });
+      onRelease?.({
+        startPosition: startPosition!,
+        currentPosition: currentPosition!,
+        changePosition: changePosition!,
+      });
       setStartPosition(undefined);
       setCurrentPosition(undefined);
       setIsDragging(false);
     }
   }, [isDragging, currentPosition, startPosition, changePosition, onRelease]);
 
-  useEventListener('mousemove', onMouseMoveWindow as any);
-  useEventListener('mouseup', onMouseUpWindow);
+  useEventListener("mousemove", onMouseMoveWindow as any);
+  useEventListener("mouseup", onMouseUpWindow);
 
   return {
     currentPosition,
